refactor(scene): split scene.c monolith into focused translation units
Split the 1193-LOC scene.c into 4 focused translation units:
- scene_init.c (476 LOC): initialization, HDR scanning, resource setup
- scene_render.c (494 LOC): GPU buffer updates, rendering pipeline
- scene_cleanup.c (121 LOC): resource teardown
- scene_nbody.c (138 LOC): N-body simulation toggle and update

Add scene_internal.h for cross-TU internal declarations (not public API).
Update CMakeLists.txt to reference the 4 new source files.
Remove 3 pre-existing NOLINT suppressions:
- misc-include-cleaner on glm_translate (transitive include sufficient)
- insecureAPI.DeprecatedOrUnsafeBufferHandling: memset → compound literal
- insecureAPI.DeprecatedOrUnsafeBufferHandling: memcpy → typed assignment

All 63 tests pass. No new warnings. No public API changes.

Closes #230
